aggiunti timestamp element per flag nuovo e foto elementi per test

+++ app/Element.php
@@ -42,12 +42,10 @@ class Element extends Model
 
     public static function deleteAll($el){
         $path = str_replace("/","\\", $el->pathPhoto);
-        unlink(public_path('storage'.$path));
+        if(file_exists($path)){
+            unlink(public_path('storage'.$path));
+        }
 
         Element::where('id', $el->id)->delete();
     }
-
-
-    public $timestamps = false;
-
 }
