+++ app/ElementsShowRoom.php
@@ -26,8 +26,9 @@ class ElementsShowRoom extends Model
 
     public static function deleteAll($el){
         $path = str_replace("/","\\", $el->pathPhoto);
-        unlink(public_path('storage'.$path));
-
+        if(file_exists($path)){
+            unlink(public_path('storage'.$path));
+        }
         ElementShowRoom::where('id', $el->id)->delete();
     }
 }
