+++ app/PhotoElement.php
@@ -21,7 +21,9 @@ class PhotoElement extends Model
         if(isset($arrayPhoto)){
             foreach($arrayPhoto as $pho) {
                 $path = str_replace("/","\\",$pho->path);
-                unlink(public_path('storage'.$path));
+                if(file_exists($path)){
+                    unlink(public_path('storage'.$path));    
+                }
             }  
         }
